NXT-112 Refactor chapters pagination and move into its own component
